+++ tf-modules/ingest/variables.tf
@@ -1,3 +1,4 @@
+
 variable "buckets" {
   type    = map(object({ name = string, type = string }))
   default = {}